+++ CHANGELOG.rst
@@ -13,6 +13,9 @@ History
 - The removal of the `TracerImp` object implies there is no longer an `imp()` method. Implementation-specific methods now exist in undifferentiated form on the tracer object.  Note: the OpenTracing library no longer forces the user to be explicit when calling implementation-specific methods; this has the downside that it is less obvious when an application becomes dependent on a specific tracing implementation.
 - `initNewTracer()` has been removed. The Tracer implementation object should be created directly by the library user.
 - The default no-op `Tracer.extract()` implementation now returns an empty SpanContext rather than null
+- `childOf` and `followFrom` have been moved from `Tracer` to standalone functions
+- `Tracer.flush` has been removed. It is not part of the OpenTracing spec.
+- Support for the `Tracer.startSpan(fields)` signature has been removed. It must always be called as `Tracer.startSpan(name, [fields])`
 
 *Non-breaking changes*
