+++ lib-debug/global_tracer.js
@@ -46,34 +46,22 @@ var GlobalTracerDelegate = function (_Tracer) {
     }
 
     _createClass(GlobalTracerDelegate, [{
-        key: '_startSpan',
-        value: function _startSpan() {
+        key: 'startSpan',
+        value: function startSpan() {
             var tracer = _globalTracer || noopTracer;
-            return tracer._startSpan.apply(tracer, arguments);
+            return tracer.startSpan.apply(tracer, arguments);
         }
     }, {
-        key: '_reference',
-        value: function _reference() {
+        key: 'inject',
+        value: function inject() {
             var tracer = _globalTracer || noopTracer;
-            return tracer._reference.apply(tracer, arguments);
+            return tracer.inject.apply(tracer, arguments);
         }
     }, {
-        key: '_inject',
-        value: function _inject() {
+        key: 'extract',
+        value: function extract() {
             var tracer = _globalTracer || noopTracer;
-            return tracer._inject.apply(tracer, arguments);
-        }
-    }, {
-        key: '_extract',
-        value: function _extract() {
-            var tracer = _globalTracer || noopTracer;
-            return tracer._extract.apply(tracer, arguments);
-        }
-    }, {
-        key: '_flush',
-        value: function _flush() {
-            var tracer = _globalTracer || noopTracer;
-            return tracer._flush.apply(tracer, arguments);
+            return tracer.extract.apply(tracer, arguments);
         }
     }]);
