+++ lib-debug/index.js
@@ -8,9 +8,13 @@ var _constants = require('./constants');
 
 var Constants = _interopRequireWildcard(_constants);
 
+var _functions = require('./functions');
+
+var Functions = _interopRequireWildcard(_functions);
+
 var _noop = require('./noop');
 
-var noop = _interopRequireWildcard(_noop);
+var Noop = _interopRequireWildcard(_noop);
 
 var _binary_carrier = require('./binary_carrier');
 
@@ -59,9 +63,9 @@ module.exports = _extend({
     SpanContext: _span_context2.default,
     Span: _span2.default,
     Tracer: _tracer2.default
-}, Constants, GlobalTracer);
+}, Constants, Functions, GlobalTracer);
 
 // Initialize the noops last to avoid a dependecy cycle between the classes.
-noop.initialize();
+Noop.initialize();
 
 //# sourceMappingURL=index.js.map
