+++ lib-debug/mock_tracer/mock_tracer.js
@@ -42,13 +42,12 @@ var MockTracer = function (_opentracing$Tracer) {
         // OpenTracing implementation
         //------------------------------------------------------------------------//
 
-        value: function _startSpan(fields) {
+        value: function _startSpan(name, fields) {
             // _allocSpan is given it's own method so that derived classes can
             // allocate any type of object they want, but not have to duplicate
             // the other common logic in startSpan().
             var span = this._allocSpan(fields);
-
-            span.setOperationName(fields.operationName);
+            span.setOperationName(name);
             this._spans.push(span);
 
             if (fields.references) {
@@ -70,14 +69,6 @@ var MockTracer = function (_opentracing$Tracer) {
         key: '_extract',
         value: function _extract(format, carrier) {
             throw new Error('NOT YET IMPLEMENTED');
-        }
-    }, {
-        key: '_flush',
-        value: function _flush(callback) {
-            this.clear();
-            if (callback) {
-                callback(null);
-            }
         }
 
         //------------------------------------------------------------------------//
