+++ lib-debug/tracer.js
@@ -16,17 +16,17 @@ var _span_context = require('./span_context');
 
 var _span_context2 = _interopRequireDefault(_span_context);
 
-var _reference2 = require('./reference');
-
-var _reference3 = _interopRequireDefault(_reference2);
-
 var _constants = require('./constants');
 
 var Constants = _interopRequireWildcard(_constants);
 
+var _functions = require('./functions');
+
+var Functions = _interopRequireWildcard(_functions);
+
 var _noop = require('./noop');
 
-var noop = _interopRequireWildcard(_noop);
+var Noop = _interopRequireWildcard(_noop);
 
 function _interopRequireWildcard(obj) { if (obj && obj.__esModule) { return obj; } else { var newObj = {}; if (obj != null) { for (var key in obj) { if (Object.prototype.hasOwnProperty.call(obj, key)) newObj[key] = obj[key]; } } newObj.default = obj; return newObj; } }
 
@@ -85,10 +85,7 @@ var Tracer = function () {
          *         reference: Tracer.childOf(parent.context()),
          *     });
          *
-         * @param {string|object} nameOrFields - if the given argument is a
-         *        string, it is the name of the operation and the second `fields`
-         *        argument is optional. If it is an object, it is treated as the
-         *        fields argument and a second argument should not be provided.
+         * @param {string} name - the name of the operation.
          * @param {object} [fields] - the fields to set on the newly created span.
          * @param {string} [fields.operationName] - the name to use for the newly
          *        created span. Required if called with a single argument.
@@ -109,57 +106,31 @@ var Tracer = function () {
          *        to represent time values with sub-millisecond accuracy.
          * @return {Span} - a new Span object.
          */
-        value: function startSpan(nameOrFields, fields) {
+        value: function startSpan(name, fields) {
             // Debug-only runtime checks on the arguments
             if (process.env.NODE_ENV === 'debug') {
                 if (arguments.length > 2) {
                     throw new Error('Invalid number of arguments.');
                 }
-                if (typeof nameOrFields !== 'string' && (typeof nameOrFields === 'undefined' ? 'undefined' : _typeof(nameOrFields)) !== 'object') {
-                    throw new Error('argument expected to be a string or object');
+                if (typeof name !== 'string') {
+                    throw new Error('argument expected to be a string');
                 }
-                if (typeof nameOrFields === 'string' && nameOrFields.length === 0) {
+                if (name.length === 0) {
                     throw new Error('operation name cannot be length zero');
                 }
-                if ((typeof nameOrFields === 'undefined' ? 'undefined' : _typeof(nameOrFields)) === 'object') {
-                    if (arguments.length !== 1) {
-                        throw new Error('Unexpected number of arguments');
-                    }
-                    if (nameOrFields === null) {
-                        throw new Error('fields should not be null');
-                    }
-                    if (!nameOrFields.operationName) {
-                        throw new Error('operationName is a required parameter');
-                    }
-                }
-            }
-
-            // Normalize the argument so the implementation is always provided
-            // an associative array of fields.
-            if (arguments.length === 1) {
-                if (typeof nameOrFields === 'string') {
-                    fields = {
-                        operationName: nameOrFields
-                    };
-                } else {
-                    fields = nameOrFields;
-                }
-            } else {
-                fields.operationName = nameOrFields;
-            }
-            if (process.env.NODE_ENV === 'debug') {
-                if (fields.childOf && fields.references) {
+                if (fields && fields.childOf && fields.references) {
                     throw new Error('At most one of `childOf` and ' + '`references` may be specified');
                 }
-                if (fields.childOf && !(fields.childOf instanceof _span2.default || fields.childOf instanceof _span_context2.default)) {
+                if (fields && fields.childOf && !(fields.childOf instanceof _span2.default || fields.childOf instanceof _span_context2.default)) {
                     throw new Error('childOf must be a Span or SpanContext instance');
                 }
             }
 
             // Convert fields.childOf to fields.references as needed.
+            fields = fields || {};
             if (fields.childOf) {
                 // Convert from a Span or a SpanContext into a Reference.
-                var childOf = this.childOf(fields.childOf);
+                var childOf = Functions.childOf(fields.childOf);
                 if (fields.references) {
                     fields.references.push(childOf);
                 } else {
@@ -167,43 +138,7 @@ var Tracer = function () {
                 }
                 delete fields.childOf;
             }
-            return this._startSpan(fields);
-        }
-
-        /**
-         * Return a new REFERENCE_CHILD_OF reference.
-         *
-         * @param {SpanContext} spanContext - the parent SpanContext instance to
-         *        reference.
-         * @return a REFERENCE_CHILD_OF reference pointing to `spanContext`
-         */
-
-    }, {
-        key: 'childOf',
-        value: function childOf(spanContext) {
-            // Allow the user to pass a Span instead of a SpanContext
-            if (spanContext instanceof _span2.default) {
-                spanContext = spanContext.context();
-            }
-            return this._reference(Constants.REFERENCE_CHILD_OF, spanContext);
-        }
-
-        /**
-         * Return a new REFERENCE_FOLLOWS_FROM reference.
-         *
-         * @param {SpanContext} spanContext - the parent SpanContext instance to
-         *        reference.
-         * @return a REFERENCE_FOLLOWS_FROM reference pointing to `spanContext`
-         */
-
-    }, {
-        key: 'followsFrom',
-        value: function followsFrom(spanContext) {
-            // Allow the user to pass a Span instead of a SpanContext
-            if (spanContext instanceof _span2.default) {
-                spanContext = spanContext.context();
-            }
-            return this._reference(Constants.REFERENCE_FOLLOWS_FROM, spanContext);
+            return this._startSpan(name, fields);
         }
 
         /**
@@ -265,7 +200,6 @@ var Tracer = function () {
             if (spanContext instanceof _span2.default) {
                 spanContext = spanContext.context();
             }
-
             return this._inject(spanContext, format, carrier);
         }
 
@@ -318,33 +252,8 @@ var Tracer = function () {
             return this._extract(format, carrier);
         }
 
-        /**
-         * Request that any buffered or in-memory data is flushed out of the process.
-         *
-         * @param {function(err: objectg)} done - optional callback function with
-         *        the signature `function(err)` that will be called as soon as the
-         *        flush completes. `err` should be null or undefined if the flush
-         *        was successful.
-         */
-
-    }, {
-        key: 'flush',
-        value: function flush(done) {
-            // Debug-only runtime checks on the arguments
-            if (process.env.NODE_ENV === 'debug') {
-                if (arguments.length > 1) {
-                    throw new Error('Invalid number of arguments');
-                }
-                if (done !== undefined && typeof done !== 'function') {
-                    throw new Error('callback expected to be a function');
-                }
-            }
-
-            this._flush(done);
-        }
-
         // ---------------------------------------------------------------------- //
-        // Methods to be implemented by derived classes
+        // Derived classes can choose to implement the below
         // ---------------------------------------------------------------------- //
 
         // NOTE: the input to this method is *always* an associative array. The
@@ -356,16 +265,8 @@ var Tracer = function () {
 
     }, {
         key: '_startSpan',
-        value: function _startSpan(fields) {
-            return noop.span;
-        }
-
-        // The default behavior returns a valid Reference of the given type
-
-    }, {
-        key: '_reference',
-        value: function _reference(type, spanContext) {
-            return new _reference3.default(type, spanContext);
+        value: function _startSpan(name, fields) {
+            return Noop.span;
         }
 
         // The default behavior is a no-op.
@@ -379,18 +280,7 @@ var Tracer = function () {
     }, {
         key: '_extract',
         value: function _extract(format, carrier) {
-            return noop.spanContext;
-        }
-
-        // The default implementation is a no-op that directly calls the callback,
-        // presuming one is provided.
-
-    }, {
-        key: '_flush',
-        value: function _flush(done) {
-            if (done) {
-                done(null);
-            }
+            return Noop.spanContext;
         }
     }]);
