+++ lib/span.js
@@ -128,7 +128,7 @@ var Span = function () {
         }
 
         /**
-         * Adds a single tag to the span.  See `AddTags()` for details.
+         * Adds a single tag to the span.  See `addTags()` for details.
          *
          * @param {string} key
          * @param {any} value
@@ -164,9 +164,9 @@ var Span = function () {
 
     }, {
         key: 'addTags',
-        value: function addTags(keyValuePairs) {
+        value: function addTags(keyValueMap) {
 
-            this._addTags(keyValuePairs);
+            this._addTags(keyValueMap);
             // Debug-only runtime checks on the arguments
 
             return this;
@@ -242,7 +242,7 @@ var Span = function () {
         }
 
         // ---------------------------------------------------------------------- //
-        // Methods to be implemented by derived classes
+        // Methods derived classes can choose to implement
         // ---------------------------------------------------------------------- //
 
         // By default returns a no-op SpanContext.
